requirements.txt.txt
streamlit
scikit-learn
joblib
pillow
numpy